telegram_bot/requirements.txt
aiohttp~=3.11.12
aiogram~=3.17.0
python-decouple~=3.8
magic-filter~=1.0.12
lxml~=5.3.1
aiogram_calendar~=0.6.0